requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
flake8==6.0.0
Markdown==3.4.1
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
